+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Fixed sugar.RecursiveRemove for remove full path
 * Removed `driver.ResultNoRows` in `internal/xsql`
 * BROKEN CHANGE in experimental topic api: producer id on start writer now is optional
 * BROKEN CHANGE in experimental topic api: remove `WithMessageGroupID` option (because not supported now)